implement login

+++ server/main.c
@@ -1,104 +1,142 @@
-#include <stdio.h>
-#include <stdlib.h>
+							/*Tic Tac Toe Game by Ishan Ankita*/
+									/*Server*/ 
 #include <ctype.h> 
 #include <sys/types.h> 
 #include <sys/socket.h> 
-#include <sys/wait.h>
 #include <netinet/in.h>
-#include <unistd.h>
 #include <signal.h> 
 #include <string.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <unistd.h>
+#include <sys/wait.h>
+#include <arpa/inet.h>
 
-void check (char playBoard[][3]);
-int newsockfd[2]; 
-int pid[2];
+int check (char playBoard[][3], int server_fd, int pid[]);
 
-void catcher(int sig) 
+int main(void) 
 { 
-	printf("Quitting\n");
-	exit(1);
-}
-
-int main() {
-	int sockfd[2], 
-		userCount = 0,
-		ctr = 1,
+	int server_fd,
+		status,
+		pid[2],
+		client_fds[2],
+		userCount=0,
+		ctr=1,
 		row = 0,
 		column = 0,
-		choice = 0,
-		status;
-	char readChoiceBuffer[1],
+		choice=0,
+		i,
+		flag;
+	char serverRead[1],
 		ans[1],
-		x[4],
-		a[2][40],
-    	playBoard[3][3] = {
+		readBuffer[10],
+		writeBuffer[2][40],
+		*users[2][3] = {
+			{ "user1", "12345", "false" },
+			{ "user2", "23456", "false" }
+		},
+    	playBoard [3][3] = {
 			{' ',' ',' '},
 			{' ',' ',' '},
 			{' ',' ',' '}
 		}; // to display the actual game status
 	
-	struct sockaddr_in server = {AF_INET, 8001, INADDR_ANY};
-	strcpy(ans,"");
-	static struct sigaction act,act2; 
-	act.sa_handler = catcher; 
-	sigfillset(&(act.sa_mask)); 
-	sigaction(SIGPIPE, &act, 0);
-	sigaction(SIGTSTP, &act, 0);
-	sigaction(SIGINT, &act, 0);
+	struct sockaddr_in server_addr, client_addr[2];
+    socklen_t client_addr_len = sizeof(client_addr[0]);
 
-	if ((sockfd[0] = socket(AF_INET, SOCK_STREAM, 0)) == -1) { 
-		perror("Socket Call Failed"); 
-		exit(1); 
+	strcpy(ans,"");
+
+	if ((server_fd = socket(AF_INET, SOCK_STREAM, 0)) == -1) { 
+		perror("Socket Call Failed");
+		exit(1);
 	}
 
-	if (bind(sockfd[0], (struct sockaddr *) &server, sizeof(struct sockaddr_in)) == -1) {
+    server_addr.sin_family = AF_INET;
+    server_addr.sin_addr.s_addr = inet_addr("127.0.0.1");
+    server_addr.sin_port = htons(8001);
+
+	if (bind(server_fd, (struct sockaddr *) &server_addr, sizeof(struct sockaddr_in)) == -1) { 
 		perror("Bind Call Failed");
 		exit(1);
 	}
 	
 	printf("Waiting for Players to join in TicTacToe Club.. \n");
 
-	strcpy(a[0],"Waiting for the other Player to join\n");
+	strcpy(writeBuffer[0],"Waiting for the other Player to join\n");
+
+	if (listen(server_fd, 5) == -1) {
+		perror("Listen Call Failed\n");
+		exit(1);
+	}
 
 	while (userCount < 2) {
-		if (listen(sockfd[0], 5) == -1) { 
-			perror("Listen Call Failed\n"); 
-			exit(1) ; 
+		client_fds[userCount] = accept(server_fd, (struct sockaddr *) &client_addr[userCount], &client_addr_len);
+
+		read(client_fds[userCount], readBuffer, sizeof(readBuffer));
+		printf("Username: %s\n", readBuffer);
+		for (i = 0; i < 2; i++) {
+			if (!strcmp(readBuffer, users[i][0])) {
+				break;
+			}
 		}
-		
-		newsockfd[userCount] = accept(sockfd[0], 0, 0);
+		if (i == 2) {
+			strcpy(writeBuffer[0], "No such user");
+			strcpy(writeBuffer[1], "0");
+			write(client_fds[userCount], writeBuffer, sizeof(writeBuffer));
+			continue;
+		}
+		read(client_fds[userCount], readBuffer, sizeof(readBuffer));
+		printf("Password: %s\n", readBuffer);
+		if (strcmp(readBuffer, users[i][1])) {
+			strcpy(writeBuffer[0], "Wrong password");
+			strcpy(writeBuffer[1], "0");
+			write(client_fds[userCount], writeBuffer, sizeof(writeBuffer));
+			continue;
+		}
+		if (!strcmp(users[i][2], "true")) {
+			strcpy(writeBuffer[0], "User has logged in");
+			strcpy(writeBuffer[1], "0");
+			write(client_fds[userCount], writeBuffer, sizeof(writeBuffer));
+			continue;
+		}
+
+		users[i][2] = "true";
 		userCount++;
 
+		printf("Number of Players who joined the game: %d\n", userCount);
+
 		if (userCount == 1) {
-			strcpy(a[1], "0");
-			write(newsockfd[0], a, sizeof(a));
-			read(newsockfd[0], x, sizeof(x));
-			pid[0] = atoi(x);
+			strcpy(writeBuffer[1],"0");
+			write(client_fds[0], writeBuffer, sizeof(writeBuffer));
+			read(client_fds[0], readBuffer, sizeof(readBuffer));
+			pid[0] = atoi(readBuffer);
 		}
-
-		printf("No. of Players who joined the Club: %d\n", userCount);
-
+	
 		if (userCount == 2) {
-			strcpy(a[0], "Let's play Tic Tac Toe!!");
-			strcpy(a[1], "1");
-			write(newsockfd[0], a, sizeof(a));
-			strcpy(a[1], "2");
-			write(newsockfd[1], a, sizeof(a));
-			read(newsockfd[1], x, sizeof(x));
-			pid[1] = atoi(x);
+			strcpy(writeBuffer[0], "Let's play Tic Tac Toe!!");
+			strcpy(writeBuffer[1], "1");
+			write(client_fds[0], writeBuffer, sizeof(writeBuffer));
+			strcpy(writeBuffer[1], "2");
+			write(client_fds[1], writeBuffer, sizeof(writeBuffer));
+
+			read(client_fds[1], readBuffer, sizeof(readBuffer));
+			pid[1] = atoi(readBuffer);
 		}
 	}
 	
-	if (!fork()) {
-		int count = 0; 		
+	if (fork() == 0) 
+	{
+		int count=0; 		
 		
-		while (count == 0) {
-			read(newsockfd[ctr], readChoiceBuffer, sizeof(readChoiceBuffer));			
-			choice = atoi(readChoiceBuffer);
+		while (count==0) 
+		{ 	
+		
+			read(client_fds[ctr], serverRead, sizeof(serverRead));			
+			choice = atoi(serverRead);
 			printf("Server side the Integer received is: %d\n",choice);
 			row = --choice/3;
 				column = choice%3;
-			playBoard[row][column] = (ctr == 0) ? 'X' : 'O';
+			playBoard[row][column] = (ctr==0)?'X':'O';
 			
 			//check(playBoard);				
 			if(ctr == 1)
@@ -106,56 +144,53 @@ int main() {
 			else
 				ctr = 1;
 			
-			write(newsockfd[ctr],playBoard,sizeof(playBoard));
-			check(playBoard);
+			write(client_fds[ctr],playBoard,sizeof(playBoard)); 
+			if (check(playBoard, server_fd, pid)) {
+				break;
+			};
 		}
-		
-		close(newsockfd[0]);
-		exit(0);
-	} 	
+
+		close(client_fds[0]);
+		exit (0);
+	}
 	wait(&status);
-	close(newsockfd[1]);
+	close(client_fds[1]);
 }
 
-void check(char playBoard[][3]) {
+int check (char playBoard[][3], int server_fd, int pid[])
+{
 	int i;
 	char key = ' ';
 
 	// Check Rows
-	for (i = 0; i < 3; i++) {
-		if (playBoard [i][0] == playBoard [i][1] && playBoard [i][0] == playBoard [i][2] && playBoard [i][0] != ' ') {
-			key = playBoard [i][0];
-		}
-	}
-
+	for (i=0; i<3;i++)
+	if (playBoard [i][0] == playBoard [i][1] && playBoard [i][0] == playBoard [i][2] && playBoard [i][0] != ' ') key = playBoard [i][0];	
 	// check Columns
-	for (i = 0; i < 3; i++) {
-		if (playBoard [0][i] == playBoard [1][i] && playBoard [0][i] == playBoard [2][i] && playBoard [0][i] != ' ') {
-			key = playBoard [0][i];
-		}
-	}
-
+	for (i=0; i<3;i++)
+	if (playBoard [0][i] == playBoard [1][i] && playBoard [0][i] == playBoard [2][i] && playBoard [0][i] != ' ') key = playBoard [0][i];
 	// Check Diagonals
-	if (playBoard [0][0] == playBoard [1][1] && playBoard [1][1] == playBoard [2][2] && playBoard [1][1] != ' ') {
-		key = playBoard [1][1];
-	}
-	if (playBoard [0][2] == playBoard [1][1] && playBoard [1][1] == playBoard [2][0] && playBoard [1][1] != ' ') {
-		key = playBoard [1][1];
+	if (playBoard [0][0] == playBoard [1][1] && playBoard [1][1] == playBoard [2][2] && playBoard [1][1] != ' ') key = playBoard [1][1];
+	if (playBoard [0][2] == playBoard [1][1] && playBoard [1][1] == playBoard [2][0] && playBoard [1][1] != ' ') key = playBoard [1][1];
+
+	if (key == ' ') {
+		return 0;
 	}
 
-	printf("%c\n", key);
+	close(server_fd);
 
-	if (key == 'X') {
+	if (key == 'X')
+	{
 		printf("Player 1 Wins\n\n");
 		kill(pid[0], SIGUSR1);
 		kill(pid[1], SIGUSR1);
-		exit(0); 
 	}
 
-	if (key == 'O') {
+	if (key == 'O')
+	{
 		printf("Player 2 Wins\n\n");
 		kill(pid[0], SIGUSR2);
 		kill(pid[1], SIGUSR2);
-		exit(0);
 	}
+
+	return 1;
 }
